Move new MOSSE test to codes_tests dir
